+++ amelethdf-c/trunk/src/amelethdf/ah5_c_link.c
@@ -97,7 +97,7 @@ char AH5_read_lnk_group (hid_t file_id, const char *path, AH5_lnk_group_t *lnk_g
         lnk_group->nb_instances = children.nb_children;
         if (children.nb_children > 0)
         {
-            lnk_group->instances = (AH5_lnk_instance_t *) malloc(children.nb_children * sizeof(AH5_lnk_instance_t));
+			lnk_group->instances = (AH5_lnk_instance_t *) malloc((size_t) children.nb_children * sizeof(AH5_lnk_instance_t));
             for (i = 0; i < children.nb_children; i++)
             {
                 strcpy(path2, path);
@@ -133,7 +133,7 @@ char AH5_read_link (hid_t file_id, AH5_link_t *link)
         link->nb_groups = children.nb_children;
         if (children.nb_children > 0)
         {
-            link->groups = (AH5_lnk_group_t *) malloc(children.nb_children * sizeof(AH5_lnk_group_t));
+			link->groups = (AH5_lnk_group_t *) malloc((size_t) children.nb_children * sizeof(AH5_lnk_group_t));
             for (i = 0; i < children.nb_children; i++)
             {
                 strcpy(path, AH5_C_LINK);
